ARM: dts: reference nodes to extend by phandles instead
Many device trees upstream have moved to memory@BASE_ADDRESS for the
memory node instead of a unit-name-less memory node. This has resulted
in breakage[1] in barebox, because some device trees feature an upstream
memory size that's not available in all variants and the barebox device
tree must override it, not create a second correct node next to the broken
one and risk of_add_memory allocating banks with bogus sizes.

To avoid this happening in future, to memory or to other nodes, extension
of nodes should happen via phandles throughout, because a phandle is assumed
to be more stable and if one is removed a compile error results instead
of creating a new node somewhere. For nodes lacking a phandle, we can
use the &{/path} syntax to declare an anonymous phandle that will fail
to compile when the upstream path changes. This is better than having
devices fail to boot at runtime.

[1]: 0d26ce1c11a1 ("ARM: Phytec phyFLEX i.MX6: delete wrong memory node")

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>