clk: sync of_clk_get_parent_name with upstream
Upstream of_clk_get_parent_name only resorts to node name as a fallback.
Instead, it uses the name supplied for registered clocks whenever
possible. On the sama5d2, this results in sckc@f8048050 and slowck,
respectively. So far, sckc@f8048050 in the parents array couldn't be
matched with any existing clock, leading to an inaccurate clock tree.

Port over the v5.8-rc1 state of the function to fix this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>