mci: omap: add new am335x and am437x SDHCI specific compatibles
For a while now, Linux has had a sdhci-omap driver for TI's OMAP SoCs
making use of the SDHCI core library. This was in addition to the
older omap-hsmmc driver, which didn't.

Linux commit 0b4edf11187 ("ARM: dts: Move am33xx and am43xx mmc nodes to
sdhci-omap driver") changes the device tree compatible of the mmc nodes
on the am33xx and am43xx SoCs away from the omap-hsmmc to the SDHCI
driver. Add the new compatibles, so we aren't broken by the change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>