kvx: exclude dtb from malloc zone
When dtb is provided by the FSBL (first stage bootloader), it might
end up in the malloc zone initialized by barebox. While the dtb is
discarded after being unflatten, if the dtb is provided right after
barebox elf, then the malloc can overwrite it while unflattening it when
allocating nodes.
To avoid that, exclude the dtb from the malloc zone by checking if it
overlaps it. If so, determine the largest zone for the allocation and
modify the memory area to use that.

Signed-off-by: Clement Leger <cleger@kalray.eu>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>