ARM: at91: dts: sama5d27-som1: adapt to upstream partition layout
Linux v5.8-rc1 adds partitions to the flash@0 node, which now overlap
with the partitions we define and resulting in errors:

  ERROR: m25p0.barebox: New partition m25p0.barebox conflicts
	with m25p0.at91bootstrap
  ERROR: m25p0.barebox-environment: New partition
  	m25p0.barebox-environment conflicts with m25p0.bootloader

We don't want to reuse those partitions as their name may and probably
will change (they have leading zeroes in the unit address).
We could use the &{/path/to/partition} syntax, but that might be broken
soon as well, as the nodes aren't inside a partitions node with a
"fixed-partitions" compatible.
The busses also don't align with the yaml schema and might be renamed.

Nuke the full flash@0 node and supply ours and avoid these problems.
The offsets of the partitions have been adjusted to coincide with the
upstream offsets for second stage bootloader and environment.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>