i2c: stm32: support new st,stm32mp15-i2c compatible
Linux v5.8-rc1 has gained support for the FastMode+ registers on the
stm32mp15, but those are incompatible between the stm32mp15 and the
stm32f7, so a new compatible was introduced. The old compatible was
replaced for the MP15 SoC, breaking I2C in barebox.
Add the new compatible to fix this.

Unlike Linux, we just use the same setup parameters as for the stm32f7.
This is ok as long we don't want to support FastMode+.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>