fs: Use iget_locked() rather than squashfs specific function
Use iget_locked() to let the core allocate the inode. This properly
initializes all fields in the new inode, especially it adds the new
inode to the list of all inodes for the filesystem. This prevents a NULL
pointer derefence when iput() removes the inode from that list.

This fixes squashfs support which is broken since 43902e5763 ("fs: free
inodes we no longer need")

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>