FIT: Parse `load` and `entry` addresses.
According to the U-Boot documentation for the FIT file format, the load
and entry have to be allways defined for a "kernel" or "standalone".
But Barebox ignored the parameters. That changes with this patch.

For backward compatibility the default values for load or entry are
still used if they are not given by the FIT file.

Signed-off-by: Christian Mauderer <christian.mauderer@embedded-brains.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>