libfile: copy_file: Fix ftruncate called on device files
ftruncate may only be called on regular files. This became broken in
a5f73a6dcd. Revert to the previous behaviour and pass O_TRUNC to the
initial open() call instead of truncating it with ftruncate().

Also we can call ftruncate() to the final size also when the file
previously didn't exist.

Fixes: a5f73a6dcd ("libfile: copy_file: explicitly truncate to final size")
Reported-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>