net: phy: Add and use driver register macro
Add driver registration macros for phy drivers similar to the existing
platform device macros. This also changes the initcall level from
fs_initcall to device_initcall for the phy drivers. It is not clear
why the phy driver have been at fs_initcall in the first place, changing
it shouldn't be a problem.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>