ARM: am335x: Enable TSC/ADC clock
Now that there is a Barebox driver for the ADC, enable the clock to the
TSC/ADC block.

This machine doesn't use common clock framework clocks in Barebox, and
instead enables any clock that might get used.  Enalbing the clock could
be made conditional based on enabling the driver, but none of the other
clocks do this.

Signed-off-by: Trent Piepho <trent.piepho@synapse.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>