ARM: at91: support sama5 low level clock setup with oscillator
AT91Bootstrap has a CONFIG_MCK_BYPASS option with following help text:

  "Use external 8 to 24 Mhz clock signal as source of main clock
  instead of an external crystal oscillator.
  This option disables the internal driving on the XOUT pin.
  The external source has to provide a stable clock on the XIN pin.
  If this option is disabled, the SoC expects a crystal oscillator
  that needs driving on both XIN and XOUT lines."

When the low level clock setup was ported over from AT91Bootstrap,
this config option was dropped. It wasn't necessary for the xplained
boards that have a crystal, but for ones with an oscillator, it's
required. Add the bit flag back.

Suggested-by: Michael Grzeschik <mgr@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>